# Tag parsing optimizations