Profiling shows that for our environment, parsing tags is the single biggest steady state contributor to CPU usage.  This PR replaces some calls to strings library functions to simpler implementations with the intention of reducing CPU overhead and some memory allocations.  

Here is an example of the before scenario:
```
$ time go test -bench=BenchmarkParseDogStatsDTagsToLabels -test.benchmem -benchtime 60s
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkParseDogStatsDTagsToLabels/1_tag_w/hash-8              300000000              319 ns/op             392 B/op          6 allocs/op
BenchmarkParseDogStatsDTagsToLabels/1_tag_w/o_hash-8            300000000              318 ns/op             392 B/op          6 allocs/op
BenchmarkParseDogStatsDTagsToLabels/2_tags,_mixed_hashes-8      200000000              490 ns/op             448 B/op          9 allocs/op
BenchmarkParseDogStatsDTagsToLabels/3_long_tags-8               100000000              673 ns/op             504 B/op         12 allocs/op
BenchmarkParseDogStatsDTagsToLabels/a-z_tags-8                  20000000              5408 ns/op            3622 B/op         57 allocs/op
PASS
ok      github.com/prometheus/statsd_exporter   584.443s
```

And here is what it looks like afterwards:

```
$ time go test -bench=BenchmarkParseDogStatsDTagsToLabels -test.benchmem -benchtime 60s
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkParseDogStatsDTagsToLabels/1_tag_w/hash-8              500000000              209 ns/op             344 B/op          4 allocs/op
BenchmarkParseDogStatsDTagsToLabels/1_tag_w/o_hash-8            500000000              201 ns/op             344 B/op          4 allocs/op
BenchmarkParseDogStatsDTagsToLabels/2_tags,_mixed_hashes-8      300000000              282 ns/op             352 B/op          6 allocs/op
BenchmarkParseDogStatsDTagsToLabels/3_long_tags-8               200000000              393 ns/op             360 B/op          8 allocs/op
BenchmarkParseDogStatsDTagsToLabels/a-z_tags-8                  30000000              2871 ns/op            2374 B/op         30 allocs/op
PASS
ok      github.com/prometheus/statsd_exporter   566.397s
```